daemon: split route/tunnel helpers out of daemon_run.go
Pure code-motion refactor toward the #5661 modularity cohort target
(daemon_run.go was ~2820 prod LOC, over the ~2000 threshold). This
increment lifts the route- and tunnel-inference helpers into a new
sibling file in package daemon:

  - riMemberLinuxName
  - collectAppliedTunnels
  - linkLocalV6Net
  - inferIPv6StaticNextHopInterfaces

Declarations are moved verbatim; no rename, no reorder, no logic
change. Only the per-file import set is narrowed to what each file
uses. buildRuntimeDataPlane is deliberately LEFT in daemon_run.go: the
retirement-boundary canary (TestDaemonRuntimeEntryPointUsesRuntime-
DataPlane) parses daemon_run.go and requires a dataplane.NewRuntime-
DataPlane call there.

Validation: go build ./..., go test ./pkg/daemon/..., and the
pkg/dataplane retirement-boundary canary all pass; gofmt clean.
